pyproject.toml
[project]
name = "tentaclepreview"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "gitpython>=3.1.45",
    "pygithub>=2.6.1",
    "requests>=2.32.4",
]
